+++ pkg/providers/appsec/testdata/TestDSAdvancedSettingsAttackPayloadLogging/match_by_id.tf
@@ -1,0 +1,10 @@
+provider "akamai" {
+  edgerc        = "../../test/edgerc"
+  cache_enabled = false
+}
+
+data "akamai_appsec_advanced_settings_attack_payload_logging" "test" {
+  config_id = 43253
+}
+
+